requirements.txt
# Core AI
openai==1.3.5
python-dotenv==1.0.0

# Data processing
pandas==2.1.3
pyyaml==6.0.1

# Utils
tqdm==4.66.1

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
psycopg2-binary==2.9.9

# HTTP requests
requests==2.31.0
urllib3==2.1.0
